Refactor client-portal-test to use extension-iframe-sdk
- Convert UI from hand-rolled postMessage to IframeBridge SDK
- Add esbuild bundling for UI TypeScript
- Rename SDK package to @alga-psa/extension-iframe-sdk
