+++ sdk/extension-iframe-sdk/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@alga/extension-iframe-sdk",
+  "name": "@alga-psa/extension-iframe-sdk",
   "version": "0.1.0",
   "private": false,
   "license": "BSD-3-Clause",
